+++ CHANGELOG.md
@@ -37,6 +37,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.1.8] - 2017-08-24
 - Added authentication URL in response details interface for generic use
-
+ 
 ## [1.1.9] - 2017-09-13
 - Changed hardcoded client base uri to optional constructor parameter.
+
+## [1.1.10] - 2017-09-13
+- Removed charge validation rule to check the usage of WireTransfer.
+- Removed charge validation rule to check currency.